Move tryGetRemoteInitialSlideFromUrl and related tests to a more reasonable place
