+++ src/shared/services/commandUtils.test.ts
@@ -255,29 +255,4 @@ RETURN n`
       })
     })
   })
-
-  describe('tryGetRemoteInitialSlideFromUrl', () => {
-    it('extracts initial slide hashbangs from a string', () => {
-      expect(utils.tryGetRemoteInitialSlideFromUrl('foo#slide-1')).toEqual(1)
-      expect(
-        utils.tryGetRemoteInitialSlideFromUrl('http://foo.com#slide-2')
-      ).toEqual(2)
-      expect(
-        utils.tryGetRemoteInitialSlideFromUrl(
-          'http://www.google.com/yarr/#slide-21'
-        )
-      ).toEqual(21)
-    })
-    it('returns 0 when no valid hashbang found', () => {
-      expect(utils.tryGetRemoteInitialSlideFromUrl('foo')).toEqual(0)
-      expect(
-        utils.tryGetRemoteInitialSlideFromUrl('http://foo.com#sloide-2')
-      ).toEqual(0)
-      expect(
-        utils.tryGetRemoteInitialSlideFromUrl(
-          'http://www.google.com/yarr/#slide-fooo'
-        )
-      ).toEqual(0)
-    })
-  })
 })
